+++ cmd/channel-manipulation/bandbool.js
@@ -34,8 +34,7 @@ const queue = require('../../lib/queue')
 const positionals = {
   operator: {
     choices: constants.BOOL,
-    desc: 'Operator to perform that bitwise operation',
-    type: 'string'
+    desc: 'Operator to perform that bitwise operation'
   }
 }