+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 4.0.0-dev
 * Updated `yargs` dependency.
+* Added [`affine`](https://sharp.pixelplumbing.com/api-operation#affine) command.
 * Removed `--density` from `joinChannel` command.
 * Renamed `--band` to `--channel` (`extractChannel` command).
 * Updated `composite` command to accept multiple images ([#73](https://github.com/vseventer/sharp-cli/issues/73)).